refactor(kwin-effect): extract nested types to plasmazoneseffect/types.h
Phase 2 of the file split. Moves six nested struct types out of the
PlasmaZonesEffect class scope into namespace-level types in PlasmaZones:
WindowBorder, CachedTexture, CachedShader, ShaderTransition,
CachedSnapRestore, EffectVirtualScreenDef. Defined in a new
plasmazoneseffect/types.h that the header includes; both the class
declaration and the sibling translation units find them via the
enclosing namespace.

No API change visible to consumers — the types kept their names; only
their qualification path changes (PlasmaZonesEffect::Foo → PlasmaZones::Foo).
All call sites used unqualified names from inside the namespace, so no
.cpp file changes were required.

plasmazoneseffect.h drops from 1196 to 963 lines.
