+++ kwin-effect/plasmazoneseffect/types.h
@@ -1,0 +1,238 @@
+// SPDX-FileCopyrightText: 2026 fuddlesworth
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+#pragma once
+
+#include <PhosphorAnimation/AnimationShaderContract.h>
+
+#include <opengl/glshader.h>
+#include <opengl/gltexture.h>
+#include <scene/borderradius.h>
+
+#include <QMetaObject>
+#include <QPointer>
+#include <QRect>
+#include <QString>
+#include <QVector4D>
+
+#include <array>
+#include <memory>
+
+namespace KWin {
+class EffectWindow;
+class Item;
+class OutlinedBorderItem;
+}
+
+namespace PlasmaZones {
+
+/// Per-window native border (scene graph item).
+/// `item` is a QPointer because OutlinedBorderItem is parented to WindowItem —
+/// Qt parent-child ownership may destroy it before removeWindowBorder() runs.
+struct WindowBorder
+{
+    QPointer<KWin::OutlinedBorderItem> item;
+    QMetaObject::Connection geometryConnection;
+    QPointer<KWin::Item> clippedContainer;
+    KWin::BorderRadius savedContainerRadius;
+};
+
+/// User-texture cache entry. Owns the uploaded `GLTexture` and tracks the wrap
+/// mode last applied to its GL state — so two shader transitions sharing the
+/// same on-disk path can run with different wrap modes without invalidating
+/// each other's cache entry, and paintWindow can skip redundant `setWrapMode`
+/// calls (each is two `glTexParameteri`s; on a high-Hz display with multiple
+/// slots in flight the redundant traffic is non-trivial).
+struct CachedTexture
+{
+    std::unique_ptr<KWin::GLTexture> texture;
+    GLenum lastAppliedWrap = GL_CLAMP_TO_EDGE;
+    /// Monotonic access counter (NOT a wall-clock frame number). Updated on
+    /// every successful lookup or fresh insert. The LRU eviction sweep
+    /// evicts the entry with the smallest value (least recently used) that
+    /// is NOT currently referenced by any in-flight transition.
+    quint64 lastAccessTick = 0;
+};
+
+/// Compiled shader + cached uniform locations for the animation contract.
+struct CachedShader
+{
+    std::unique_ptr<KWin::GLShader> shader;
+    /// Animation-shader contract uniform locations (see
+    /// `PhosphorAnimationShaders::AnimationShaderContract`). Per-effect
+    /// declared parameters land in the `customParams[N]` slots — the
+    /// translation from friendly parameter ids (e.g. `direction`) to slot
+    /// keys is performed by `AnimationShaderRegistry::translateAnimationParams`
+    /// when the transition starts; paintWindow just pushes vec4s.
+    int iTimeLoc = -1;
+    int iResolutionLoc = -1;
+    int iTimeDeltaLoc = -1;
+    int iFrameLoc = -1;
+    int iDateLoc = -1;
+    int iMouseLoc = -1;
+    int iIsReversedLoc = -1;
+    // Slot counts sourced from AnimationShaderContract so a future change to
+    // the contract (e.g. growing the customParams budget) can't silently
+    // desync this cache from the translation + upload sites in
+    // shader_transitions.cpp. The default-initialiser sets every entry to -1;
+    // std::array's value-initialisation doesn't, so wrap the construction.
+    std::array<int, PhosphorAnimationShaders::AnimationShaderContract::kMaxCustomParams> customParamsLoc = []() {
+        std::array<int, PhosphorAnimationShaders::AnimationShaderContract::kMaxCustomParams> a;
+        a.fill(-1);
+        return a;
+    }();
+    std::array<int, PhosphorAnimationShaders::AnimationShaderContract::kMaxCustomColors> customColorsLoc = []() {
+        std::array<int, PhosphorAnimationShaders::AnimationShaderContract::kMaxCustomColors> a;
+        a.fill(-1);
+        return a;
+    }();
+    /// Sampler uniform locations for `uTexture1..3`. -1 means the GLSL linker
+    /// dropped the sampler (the shader source never read it), in which case
+    /// paintWindow skips both the bind and the setUniform — saves a
+    /// glActiveTexture call per unused slot.
+    std::array<int, PhosphorAnimationShaders::AnimationShaderContract::kMaxUserTextureSlots> userTextureLoc = []() {
+        std::array<int, PhosphorAnimationShaders::AnimationShaderContract::kMaxUserTextureSlots> a;
+        a.fill(-1);
+        return a;
+    }();
+    /// Element locations for `iTextureResolution[0..N-1]`. Symmetric with
+    /// `customParamsLoc` / `customColorsLoc` — the element-name lookup
+    /// happens at compile time so paintWindow does not re-resolve
+    /// `"iTextureResolution[0]"` string lookups per frame.
+    std::array<int, PhosphorAnimationShaders::AnimationShaderContract::kMaxUserTextureSlots> iTextureResolutionLoc =
+        []() {
+            std::array<int, PhosphorAnimationShaders::AnimationShaderContract::kMaxUserTextureSlots> a;
+            a.fill(-1);
+            return a;
+        }();
+};
+
+/// Per-window in-flight shader transition.
+struct ShaderTransition
+{
+    const CachedShader* cached = nullptr;
+    /// `customParams[N]` slot values resolved at transition begin time.
+    /// `AnimationShaderRegistry::translateAnimationParams` translates the
+    /// friendly parameter map (e.g. `{"direction": 1, "parallax": 0.2}`) to
+    /// slot keys (`{"customParams1_x": 1, "customParams1_y": 0.2}`); we pack
+    /// those into vec4s here so paintWindow only does 8 setUniform calls
+    /// per frame, not per-frame string lookups. Slots with no declared
+    /// parameters stay at (0, 0, 0, 0).
+    std::array<QVector4D, PhosphorAnimationShaders::AnimationShaderContract::kMaxCustomParams> customParamsValues = {};
+    std::array<QVector4D, PhosphorAnimationShaders::AnimationShaderContract::kMaxCustomColors> customColorsValues = {};
+    /// Two-mode progress source.
+    /// • `durationMs > 0`: time-based — `startTimeMs` is the monotonic
+    ///   `shaderClockNowMs()` (steady_clock) at begin time and paintWindow
+    ///   computes `progress = clamp01((now - startTimeMs) / durationMs)`.
+    ///   Used by lifecycle events (window.open/close/focus/etc.) that have
+    ///   no `m_windowAnimator` animation to ride.
+    /// • `durationMs == 0`: animator-driven — paintWindow reads progress
+    ///   from `m_windowAnimator->animationFor(w)->state().value`. Used by
+    ///   zone.* events that flow through `applySnapGeometry` and inherit
+    ///   the geometry animation's timeline.
+    qint64 startTimeMs = 0;
+    int durationMs = 0;
+    /// Monotonic per-window generation. Each `beginShaderTransition` bumps
+    /// the counter for that window; the timer-driven `endShaderTransition`
+    /// scheduled by `tryBeginShaderForEvent` captures the generation at
+    /// schedule time and bails on fire if the live transition has been
+    /// superseded — without this, a stale timer would tear down a fresh
+    /// transition that replaced it before it had a chance to play out.
+    quint64 generation = 0;
+    /// When true, paintWindow flips progress to 1 - progress before the
+    /// uniform write, so the shader runs from `iTime = 1.0` down to
+    /// `iTime = 0.0` over the transition's lifetime. Used by lifecycle
+    /// events whose semantic is "going away" — window.close,
+    /// going-to-minimized, going-to-unmaximized — so a single shader effect
+    /// doubles as both directions: an `appear` shader (e.g. fade-in from
+    /// glitch) reads progress=0 as "fully obscured" and progress=1 as
+    /// "fully revealed", which is exactly the semantic users expect for
+    /// the corresponding `disappear` event.
+    bool reverse = false;
+    /// Per-leg frame counter. Bumped each paintWindow tick where this
+    /// transition feeds the shader; reset to 0 on every fresh
+    /// beginShaderTransition install (or supersession). Mirrors the daemon's
+    /// `SurfaceAnimator` `iFrame` semantic: starts at 0 on each fresh
+    /// attach so shaders that read it (e.g. for staggered per-frame
+    /// randomness) see the same trajectory on both runtimes.
+    ///
+    /// Type pinned to `int` to match GLSL `uniform int iFrame;` in
+    /// `data/animations/shared/animation_uniforms.glsl`. The shared
+    /// header's UBO offset (76) is also `int`-sized so the kwin
+    /// `setUniform(iFrameLoc, iFrameValue)` push and the daemon std140
+    /// layout agree. A static_assert at the push site pins this contract.
+    int frameCount = 0;
+    /// Per-leg user-texture pointers. Resolved at `beginShaderTransition`
+    /// time from the translated params' `uTexture<N>` keys (pack defaults
+    /// merged with any per-leg runtime override). Raw non-owning pointers
+    /// into `m_textureCache`, which owns the underlying `GLTexture` and
+    /// outlives every transition that references it. nullptr means
+    /// "no texture for this slot" — paintWindow skips the bind and the
+    /// sampler reads transparent black. Pointing at the cache entry rather
+    /// than the bare `GLTexture` lets paintWindow skip redundant
+    /// `setWrapMode` calls by comparing against
+    /// `CachedTexture::lastAppliedWrap`.
+    std::array<CachedTexture*, PhosphorAnimationShaders::AnimationShaderContract::kMaxUserTextureSlots> userTextures =
+        {};
+    /// Wrap-mode GL enum applied at bind time (GL_CLAMP_TO_EDGE / GL_REPEAT
+    /// / GL_MIRRORED_REPEAT). Stored per-leg rather than on the cached
+    /// `GLTexture` so two transitions sharing the same on-disk path can
+    /// run with different wrap modes without invalidating each other's
+    /// cache entry.
+    std::array<GLenum, PhosphorAnimationShaders::AnimationShaderContract::kMaxUserTextureSlots> userTextureWrap = []() {
+        std::array<GLenum, PhosphorAnimationShaders::AnimationShaderContract::kMaxUserTextureSlots> a;
+        a.fill(GL_CLAMP_TO_EDGE);
+        return a;
+    }();
+    /// Wall-clock timestamp of the previous paintWindow tick that fed this
+    /// transition. -1 means "no prior paint yet" — first paint produces
+    /// `iTimeDelta = 0` rather than a spurious huge delta from the install
+    /// timestamp. `shaderClockNowMs()` based, so monotonic and immune to
+    /// NTP jumps mid-transition.
+    qint64 lastPaintTimeMs = -1;
+    /// True when this transition holds @c KWin::WindowClosedGrabRole on the
+    /// window. The grab keeps a closing window alive past KWin's normal
+    /// unmap-and-delete sequence so paintWindow has frames to run the
+    /// close shader on. OffscreenEffect's @c redirect alone is
+    /// insufficient — the OffscreenEffect docstring explicitly states
+    /// "The window will be automatically unredirected if it's deleted",
+    /// meaning a closing window is auto-released without the grab and the
+    /// close shader never gets a paint cycle.
+    ///
+    /// Set true only by the @c slotWindowClosed → @c tryBeginShaderForEvent
+    /// path (via the @c holdCloseGrab parameter threaded through
+    /// @c beginShaderTransition). Released in @c endShaderTransition
+    /// unconditionally — clearing the role on a non-closing window is a
+    /// no-op, and clearing it on a deleted window lets KWin proceed with
+    /// final destruction.
+    bool closeGrabHeld = false;
+};
+
+/// Pre-computed snap restore target for a pending app
+/// (appId → geometry + saved screen).
+/// Fetched once from daemon on ready; consumed in slotWindowAdded for
+/// instant teleport (no D-Bus round-trip visible flash). The screenId
+/// lets the effect tell "cached saved zone is on snap-mode screen X" from
+/// "current KWin placement is autotile screen Y" — we trust the saved
+/// screen, not the placement, so cross-VS / cross-monitor restores work.
+struct CachedSnapRestore
+{
+    QRect geometry;
+    QString screenId;
+};
+
+/// A single virtual screen subdivision within a physical monitor.
+///
+/// Virtual screens divide a physical monitor into independent sub-screens,
+/// each with its own zones, autotile state, etc. The daemon manages
+/// definitions; the effect fetches them via D-Bus and resolves positions.
+///
+/// Named `EffectVirtualScreenDef` to avoid collision with the daemon's
+/// `Phosphor::Screens::VirtualScreenDef` (which has many more fields).
+struct EffectVirtualScreenDef
+{
+    QString id; ///< e.g., "Dell:U2722D:115107/vs:0"
+    QRect geometry; ///< Absolute geometry in global compositor coords
+};
+
+} // namespace PlasmaZones
